+++ utils/collections.go
@@ -6,12 +6,12 @@ func Unshift[T any](slice []T, v T) []T {
 	return slice
 }
 
-type callback[I any, O any] func(input I, index int) O
+type callback[I any, O any] func(input I, index uint) O
 
 func Map[I any, O any](list []I, callback callback[I, O]) []O {
 	l := make([]O, len(list))
 	for i, v := range list {
-		l[i] = callback(v, i)
+		l[i] = callback(v, uint(i))
 	}
 	return l
 }
